+++ pyproject.toml
@@ -109,6 +109,7 @@ line-length = 88
 show-fixes = true
 exclude = [
   "weave/api.py",
+  "weave/query_api.py",
   "weave/__init__.py",
   "weave/legacy/**/*.py",
   "examples",
